Make BaseProvider a concrete OpenAI-compatible provider
Every provider already inherits BaseProvider, so the OpenAI wire format belongs there rather than in a separate OpenAICompatibleProvider, now removed.
